--- .dockerignore
@@ -1,2 +1,0 @@
-.git
-.idea
