--- bin/run.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-go run cmd/example_app/main.go -conf config.yml
